requirements.txt
certifi==2017.11.5
dj-database-url==0.4.2
Django==1.11.7
django-environ==0.4.4
djangorestframework==3.7.3
gunicorn==19.7.1
psycopg2==2.7.3.2
pytz==2017.3
six==1.11.0
whitenoise==3.3.1
